Jinsu

+++ src/main/java/com/capstone/ai_painter_backen/service/mentor/CategoryService.java
@@ -1,7 +1,9 @@
 package com.capstone.ai_painter_backen.service.mentor;
 
 import com.capstone.ai_painter_backen.domain.mentor.CategoryEntity;
+import com.capstone.ai_painter_backen.exception.BusinessLogicException;
 import com.capstone.ai_painter_backen.exception.DuplicateNameException;
+import com.capstone.ai_painter_backen.exception.ExceptionCode;
 import com.capstone.ai_painter_backen.mapper.etc.CategoryMapper;
 import com.capstone.ai_painter_backen.repository.mentor.CategoryRepository;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +30,7 @@ public class CategoryService {
         CategoryEntity category = categoryMapper.requestSaveDtoToEntity(requestSaveDto);
 
         if (categoryRepository.existsByCategoryName(category.getCategoryName())) {
-            throw new DuplicateNameException();
+            throw new BusinessLogicException(ExceptionCode.DUPLICATE_CATEGORY_NAME);
         }
 
         return categoryRepository.save(category).getId();
@@ -38,29 +40,47 @@ public class CategoryService {
     @Transactional
     public ResponseDto updateCategory(Long id, RequestUpdateDto requestUpdateDto) {
         CategoryEntity category = categoryRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException());
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.NO_SUCH_CATEGORY));
+
+        if (categoryRepository.existsByCategoryName(requestUpdateDto.getCategoryName())) {
+            throw new BusinessLogicException(ExceptionCode.DUPLICATE_CATEGORY_NAME);
+        }
+
         category.update(requestUpdateDto.getCategoryName());
 
         return categoryMapper.entityToResponseDto(category);
     }
 
+    /*
+    카테고리 삭제하면 해당하는 Tutor는 어떻게 할까?
+    1. 튜터도 같이 삭제
+    2. default 카테고리를 하나 만들고 거기다가 집어넣기
+    화요일에 회의 때 물어보기
+     */
     @Transactional
     public void deleteCategory(RequestDeleteDto deleteDto) {
-        categoryRepository.deleteById(deleteDto.getCategoryId());
+        if(categoryRepository.existsById(deleteDto.getCategoryId())) {
+            categoryRepository.deleteById(deleteDto.getCategoryId());
+        } else {
+            throw new BusinessLogicException(ExceptionCode.NO_SUCH_CATEGORY);
+        }
     }
 
 
     public ResponseDto findCategoryById(Long id) {
         CategoryEntity category = categoryRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException());
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.NO_SUCH_CATEGORY));
 
         return categoryMapper.entityToResponseDto(category);
     }
 
-
+    /*
+    categoryResponseDto에 보면 List<CategoryTutorEntity>가 있는데 이거 삭제하는 게 좋을 것 같음.
+    일단 주석처리 해놨음.
+     */
     public List<ResponseDto> findAllCategory() {
         return categoryRepository.findAll().stream()
-                .map(m -> new ResponseDto(m.getId(), m.getCategoryName(), m.getCategoryTutorEntities()))
+                .map(m -> new ResponseDto(m.getId(), m.getCategoryName()))
                 .collect(Collectors.toList());
     }
 }
